% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alplakes_call20250415}
\alias{alplakes_call20250415}
\title{Alplake API call}
\format{
## `alplakes_call20250415`
A list with 6 items:
\describe{
  \item{time}{List of time}
  \item{lat, lng}{Latitute and Longitude of the call}
  \item{depth}{List with information on the depths}
  \item{distance}{Distance from the call}
  \item{variables}{List with the variables requested}

}
}
\usage{
alplakes_call20250415
}
\description{
Calto Alplakes API with the `get_data_from_alplakes` function with "2025-04-15" as date.
}
\keyword{datasets}
