% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alplakes_call20250528}
\alias{alplakes_call20250528}
\title{Alplake API call}
\format{
## `alplakes_call20250528`
A list with 6 items:
\describe{
  \item{time}{List of time}
  \item{lat, lng}{Latitute and Longitude of the call}
  \item{depth}{List with information on the depths}
  \item{distance}{Distance from the call}
  \item{variables}{List with the variables requested}

}
}
\usage{
alplakes_call20250528
}
\description{
Call to Alplakes API with the `get_data_from_alplakes` function with arguments
day= "2025-05-28" and end_day = "2025-05-31"
}
\keyword{datasets}
