% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{convert_json_to_list}
\alias{convert_json_to_list}
\title{convert_json_to_list}
\usage{
convert_json_to_list(API_response)
}
\arguments{
\item{API_response}{response from get_data_from_alplakes}
}
\value{
The API response read with jsonlite::fromJSON. If the call failed, it will return the
data from the 14th of June 2024.
}
\description{
convert_json_to_list
}
